Починить ошибку компиляции сводки кубатурника: DoubleArray[i] += Int
Kotlin не резолвит составное присваивание DoubleArray[i] += totalCount,
когда totalCount — Int ("No set method providing array access", нашла
реальная сборка в CI). Явное totalCount.toDouble() решает — accVolume-часть
(Double += Double) собиралась и без этого нормально.

+++ app/src/main/java/com/lesovod/mobile/ui/kubaturnik/KubaturnikViewModel.kt
@@ -143,10 +143,10 @@ data class KubaturnikUiState(
                 tables += SortDestinationTable(sort, dest, groups, totalCount, totalVolume)
 
                 val sortAcc = bySortAcc.getOrPut(sort) { DoubleArray(2) }
-                sortAcc[0] += totalCount
+                sortAcc[0] += totalCount.toDouble()
                 sortAcc[1] += totalVolume
                 val destAcc = byDestAcc.getOrPut(dest) { DoubleArray(2) }
-                destAcc[0] += totalCount
+                destAcc[0] += totalCount.toDouble()
                 destAcc[1] += totalVolume
                 grandCount += totalCount
                 grandVolume += totalVolume
